refactor: move thread runtime event foundation
